feat: move example tree visiting
