+++ src/logic/prepare/mock-map.ts
@@ -72,5 +72,11 @@ export async function prepareMockMap(
       },
       NodeFileSystem
     );
+
+    await OutputStream.writeOnce(
+      superJsonPath,
+      JSON.stringify(superJson, undefined, 2)
+    );
+    logger.info('updateSuperJson', superJsonPath);
   }
 }
